move Button component into its own folder
