+++ src/components/Modal.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import ReactDOM from 'react-dom';
-import Button from './Button';
+import Button from './Button/Button';
 import Card from './Card';
 
 interface ModalProps {
