+++ src/pages/ResultPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header';
 import Card from '../components/Card';
-import Button from '../components/Button';
+import Button from '../components/Button/Button';
 import { useResults } from '../hooks/useResults';
 import type { AnswerHistory } from '../types';
