+++ src/pages/StartPage.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header';
-import Button from '../components/Button';
+import Button from '../components/Button/Button';
 import Card from '../components/Card';
 import SettingsForm from '../components/SettingsForm';
 import Modal from '../components/Modal';
